pyproject.toml
[tool.poetry]
name = "robot-payload-id"
version = "0.1.0"
description = "Identifying object inertial properties using a robot arm"
authors = ["Nicholas Pfaff <nepfaff@mit.edu>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<3.12"
numpy = "^1.25.2"
drake = "^1.21.0"
manipulation = {git = "https://github.com/RussTedrake/manipulation.git"}
tqdm = "^4.66.1"
hydra-core = "^1.3.2"
omegaconf = "^2.3.0"

[tool.poetry.group.dev.dependencies]
pre-commit = "^3.4.0"
black = "^23.9.1"
isort = "^5.12.0"
poetry-pre-commit-plugin = "^0.1.2"

[tool.isort]
profile = 'black'
lines_between_types = 1
combine_as_imports = true
known_first_party = ['robot_payload_id']

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
